requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
requests
pandas
openpyxl
xlrd
python-dotenv